Organize integration tests helpers.
