+++ tests/integration/test_charm.py
@@ -7,15 +7,14 @@ from pathlib import Path
 
 import jubilant
 from cassandra.cluster import ResultSet
-from helpers import (
+from helpers.cassandra import (
     assert_rows,
     connect_cql,
-    get_leader_unit,
-    get_non_leader_units,
     prepare_keyspace_and_table,
     read_n_rows,
     write_n_rows,
 )
+from helpers.juju import get_leader_unit, get_non_leader_units
 
 logger = logging.getLogger(__name__)
