+++ tests/integration/test_multinode.py
@@ -7,7 +7,7 @@ from pathlib import Path
 
 import jubilant
 from cassandra.cluster import ResultSet
-from helpers import connect_cql
+from helpers.cassandra import connect_cql
 
 logger = logging.getLogger(__name__)
